ver 0.0.0.3
+
The scene was updated, the PS1 filter was put on the camera, the script (camera) was fixed, the start menu was created.
The first attempt to create a multiplayer was made +- successfully
-
Fix it :
Grab it via web mode does not work.
The multiplayer is not finalized, it is out of sync at least with the camera to check the movement.
//
Plans for the next versions
Fix bugs, make a full-fledged lobby out of the "start game" scene. Make an adequate UI and ideally make another scene with gameplay