ver 0.0.0.4
The main menu (test) has been added, the player's menu has not been completed, the tasks are to fix bugs from previous versions, namely :
-
Fix it :
"Grab it" via web mode does not work.
The multiplayer is not finalized, it is out of sync at least with the camera to check the movement.
//
Plans for the next versions
Fix bugs, make a full-fledged lobby out of the "start game" scene. Make an adequate UI and ideally make another scene with gameplay